Extract storage proposal building logic to a new class
