+++ src/Makefile.am
@@ -72,7 +72,8 @@ ynclude_DATA = \
 ylibdir = @ylibdir@/autoinstall
 ylib_DATA = \
   lib/autoinstall/module_config_builder.rb \
-  lib/autoinstall/pkg_gpg_check_handler.rb
+  lib/autoinstall/pkg_gpg_check_handler.rb \
+  lib/autoinstall/storage_proposal.rb
 
 scrconf_DATA = \
   scrconf/cfg_autoinstall.scr \
