+++ test/autoinst_storage_test.rb
@@ -7,80 +7,29 @@ describe Yast::AutoinstStorage do
   subject { Yast::AutoinstStorage }
 
   describe "#Import" do
-    let(:proposal_settings) { double("proposal_settings") }
-    let(:guided_proposal) { instance_double(Y2Storage::GuidedProposal, propose: nil, proposed?: true) }
-    let(:auto_inst_proposal) { instance_double(Y2Storage::AutoInstProposal, propose: nil, proposed?: true) }
-    let(:devicegraph) { instance_double(Y2Storage::Devicegraph) }
-    let(:storage_manager) { double("storage_manager", y2storage_probed: devicegraph) }
-    let(:disk_analyzer) { instance_double(Y2Storage::DiskAnalyzer) }
+    let(:storage_proposal) { instance_double(Y2Autoinstallation::StorageProposal) }
 
     before do
-      allow(Y2Storage::StorageManager).to receive(:instance)
-        .and_return(storage_manager)
-      allow(Y2Storage::GuidedProposal).to receive(:new)
-        .and_return(guided_proposal)
-      allow(Y2Storage::AutoInstProposal).to receive(:new)
-        .and_return(auto_inst_proposal)
-      allow(storage_manager).to receive(:proposal=)
+      allow(Y2Autoinstallation::StorageProposal).to receive(:new)
+        .and_return(storage_proposal)
     end
 
-    context "when no partitioning plan is given" do
-      let(:profile) { nil }
-
-      it "sets a proposal" do
-        expect(storage_manager).to receive(:proposal=).with(guided_proposal)
-        subject.Import(profile)
-      end
-
-      it "returns true" do
-        expect(subject.Import(profile)).to eq(true)
-      end
-
-      context "if proposal fails" do
-        before do
-          allow(guided_proposal).to receive(:propose).and_raise(Y2Storage::Error)
-        end
-
-        it "does not set the proposal" do
-          expect(storage_manager).to_not receive(:proposal=)
-          subject.Import(profile)
-        end
-
-        it "returns false" do
-          expect(storage_manager).to_not receive(:proposal=)
-          subject.Import(profile)
-        end
-      end
+    it "creates a proposal" do
+      expect(storage_proposal).to receive(:propose_and_store)
+        .and_return(true)
+      subject.Import({})
     end
 
-    context "when profile contains an empty set of partitions" do
-      let(:profile) { [] }
-
-      it "sets a proposal" do
-        expect(storage_manager).to receive(:proposal=).with(guided_proposal)
-        subject.Import(profile)
-      end
+    it "returns true if proposal succeeds" do
+      allow(storage_proposal).to receive(:propose_and_store)
+        .and_return(true)
+      expect(subject.Import({})).to eq(true)
     end
 
-    context "when a partition plan is given" do
-      let(:profile) { [{ "device" => "/dev/sda" }] }
-
-      before do
-        allow(Y2Storage::AutoInstProposal).to receive(:new).and_return(auto_inst_proposal)
-        allow(Y2Storage::DiskAnalyzer).to receive(:new).with(devicegraph).and_return(disk_analyzer)
-      end
-
-      it "asks for a proposal" do
-        expect(Y2Storage::AutoInstProposal).to receive(:new)
-          .with(partitioning: profile, devicegraph: devicegraph, disk_analyzer: disk_analyzer)
-          .and_return(auto_inst_proposal)
-        expect(auto_inst_proposal).to receive(:propose)
-        subject.Import(profile)
-      end
-
-      it "returns true" do
-        expect(subject.Import(profile)).to eq(true)
-      end
+    it "returns false if proposal fails" do
+      allow(storage_proposal).to receive(:propose_and_store)
+        .and_return(false)
+      expect(subject.Import({})).to eq(false)
     end
   end
